Settings main tab implemented with statistics and acl option, provider statistics added
