+++ src/scss/settings/stats.scss
@@ -1,11 +1,28 @@
 .content__stats{
-	height: 200px;
+	height: 100%;
     font-size: 10pt;
+
+    h4 {
+    	margin: 20px 0;
+    }
 
 	.no-stats {
         width: 100%;
         height: 70px;
         margin-top: 20px;
         text-align: center;
+    }
+
+    .statistics__task {
+    	display: grid;
+    	grid-template-columns: 50% 50%;
+    	grid-template-rows: auto;
+    	grid-row-gap: 25px;
+    	font-size: 9pt;
+    	div {
+    		span + span {
+    			margin-left: 5px;
+    		}
+    	}
     }
 }
